feat(filedropper): make S3 bucket names optional (#1)
Our forwarder can push data from SNS and Eventbridge to Filedrop as well
as S3 files. As such, we shouldn't require a list of S3 buckets.

Opening this as a pull request in order to start figuring out what we'd
hope testing would look like. What would be an appropriate test to
submit alongside this change? What tests would be run for this small
diff?