fix: support wildcard in bucket name
We want to allow wildcarding a bucket name, (e.g. "*" or "observe-*").
To do so, we need to suffix resources in the policy that grants
permission for GetObject with `/*` rather than `*`.

The current implementation is broken, in that a bucket name of `observe`
would lead to the following policy statement:

```
{
    "Action": [
        "s3:GetObject",
        "s3:GetObjectTagging"
    ],
    "Resource": [
        "arn:aws:s3:::observe*"
    ],
    "Effect": "Allow"
}
```

This would match all objects in buckets with `observe` as a prefix!