ci: fix tests
Restructure tests. This now runs reliabily for me locally, but there is still plenty of potential for race conditions.

The core change was to restructure things so that subscriptions are set
up via terraform. The shell script now is solely responsible for writing
and reading objects. I don't know why this fixed a source of flakiness,
but it appears that reusing AWS sessions helps a bit in reducing
non-determinism.

Our SNS -> SQS subscription never worked for wildcards. I've added a
TODO comment to address this later.