feat: add structured logging
We were previously using the log package as a backend to `logr`. Since
1.21, go supports `slog`, which provides native structured logging.

This commit swaps out our `logr` backend to use `slog`. We still want to
use `logr`, since it provides a nicer interface imo. There are some
compatibility quirks, in particular around log levels. `logr` expects
levels of verbosity, whereas in `slog` levels are tied to severity. For
now we just negate it.