feat: add subscriber scaffolding (#59)
* fix: use mux for forwarder, break out test package

This is a small refactor to use the handler mux in the forwarder and
break out the mock utilitiy into a separate package.

* feat: add cloudwatch logs dependency

Add small mock package to pull dependency in.

* feat: add subscriber scaffolding

This adds the basic scaffolding for a cloudwatch logs subscriber:

- a cloudformation template that sets up a lambda and an SQS queue
- a lambda function that can be triggered with a custom event
- a test that verifies triggering the lambda works

We currently don't do much in our lambda function. Given a custom event,
we paginate through log groups and write each page to the queue in order
to fan out. Since we have no processing associated to these messages,
the subsequent lambdas will error out.