fix(subscriber): return stats in response (#82)
We don't have much use for the lambda invocation response outside of
testing. For introspection purposes, it's useful to return some
information about what exactly happened. A set of stats counter is
likely the simplest and most flexible we can get away with.

These stats must be thread safe, since we'll need to eventually spawn
goroutines to do work concurrently.