feat(collection): add subscriber stack
This commit adds the subscriber stack as part of collection. By default
nothing is collected. Users can provide either a list of log prefixes or
patterns to collect, and can use the wildcard to collect all log groups.

Note that we still won't have the full "out of the box" experience yet
as far as cloudwatch logs is concerned. The subscriber is still missing
the eventbridge rules which trigger subscription.