feat(subscriber): introduce wildcard pattern
Previously we were treating the absence of patterns or prefixes to
signify "subscribe all". Semantically this is quite confusing once you
try to wire it into cloudformation, where we would like to have the
presence of inputs gate the installation of our app.

This commit adds the ability to support wildcard on either
`logGroupNamePatterns` or `logGroupNamePrefixes`. The subscriber
cloudformation template is adjusted to still subscribe all log groups on
install by default.