feat(collection): export log groups (#104)
* fix(subscriber): correct ARN validation for IAM role

Our logic was busted. It was passing tests because the test fixture had
a lambda ARN instead. Provide fix and add a new test for the happy path
of a firehose ARN alongside the role ARN.

* feat(collection): export log groups

Add static subscriptions for the log groups we do create. This should
not clash with the subscriber module since the filter name is set to
match (`observe-logs-subscription`).