fix(subscriber): use iam:PassRole for Lambda (#135)
We need to allow iam:PassRole for the subscription lambda for the case
where the destination is a Firehose Delivery Stream. We currently do not
adequately capture this functionality because we never test Firehose as
a destination. We'll have to fix that with further integration testing
in the collection stack.