fix(subscriber): set role ARN (#134)
* fix(subscriber): use iam:PassRole for Lambda

We need to allow iam:PassRole for the subscription lambda for the case
where the destination is a Firehose Delivery Stream. We currently do not
adequately capture this functionality because we never test Firehose as
a destination. We'll have to fix that with further integration testing
in the collection stack.

* fix(subscriber): set role ARN

We were not correctly configuring the role ARN when setting a subscription filter. This is only necessary when the destination is a Kinesis Firehose delivery stream. Added unit test to verify the role ARN is passed to cloudwatchlogs:PutSubscriptionFilter