feat(collection): enable cloudwatch logs discovery (#136)
Configure daily discovery event. This can be hardcoded to reduce
configuration fatigue.

As a result of this change, we can no longer explicitly configure
subscription filters, since there is a possibility of race condition
between "discovery" and the explicit filter. If a user wants to
forward the log groups created by the stack, they would have to provide
the stack name as one of the LogGroupPatterns.