fix: test name override
This commit tests `NameOverride`.

Typically we rely on Cloudformation to pick resource names for us. In
some cases however we need stable identifiers, for example IAM role
names that must be assumed from our end, or functions for which we need
to create a log group upfront.

In those cases, we can't necessarily rely on the stack name.
Cloudformation names have a maximum length of 128 characters, which far
exceeds the name limits for most resources in AWS. We need to provide an
override for passing a name that abides by the limits of all the
resources present in a given template.

This change ended up being a bit involved because I had to change the
interface of our setup module. We now create three variables on setup:

- a `stack_name` which is 128 characters long
- an `id` which is 64 characters long
- a `short` identifier which is 32 characters long

We then pass in these variables as necessary to exercise the limits of
our templates.