fix: rename subscriber to logwriter (#145)
Our subscriber app no longer subscribes log groups necessarily. We now
install a firehose by default, and only subscribe log groups if users
provide a set of patterns or prefixes to manage.

The reason for this change is simple: users may want to subscribe log
groups themselves. In that case, we will need to provide a standalone
firehose which can be written to, at which point the "subscriber" name
no longer makes any sense.