fix(config): move wildcarding to config app
This commit makes the behavior of `IncludeResourceTypes` and
`ExcludeResourceTypes` consistent across apps.

Previously we implemented "wildcard" support in the collection module.
This could result in potential confusion, since configuring the config
app as a standalone would take the same parameters as the collection
app, but behave differently. This commit moves the logic for supporting
the wildcard into the config app.

In the process we also add the `IncludeGlobalResourceTypes` parameter
for consistency with the terraform module.