feat: add metric stream filtering
This commit adds the ability to filter metrics collected by CloudWatch
Metrics Stream. This feature was sorely missing in our previous stacks,
leading to excessive collection costs.

The difficulty in filtering metrics in CloudFormation derives from its
complete lack of expressiveness. There's no way to take a user input in
the supported types and fill out the complex structure that defines
include and exclude filters.

The solution is to instead host a template snippet in S3, and allow
customers to point at their own derived versions. This allows us to
maintain a trimmed down set of metrics without affecting a customer's
ability to fine tune the set for their own needs.