refactor: formalize overrides (#183)
Formalize the concept of an override when copying an S3 object.

Previously we had a method by which users could pass a comma delimited
list of key values that would affect how the content type would be set.
We know from prior experience that over time the type of overrides will
tend to grow (e.g. ignore file, add other metadata), so it's worth
adding some structure.

This commit introduces the concept of override "sets", each set of which
is a sequence of rules. A rule allows you to match on a given copy
operation (e.g. keys ending in `.json`), and apply an action
(e.g. set content type to `application/json`).

We continue to construct a custom set from the `CONTENT_TYPE_OVERRIDES`
environment variable. We introduce a new `PRESET_OVERRIDES` env var which
references the preconfigured sets we should load. This allows us to ship
predefined rules, but still allow users to unload them if necessary.