fix: override cloudwatch metrics, adjust prefixes
This commit adds a preset for identifying cloudwatch metrics stream
data, which will be submitted with content type
`application/x-aws-cloudwatchmetrics`.

In the process, I've how object prefixes get set. Previously we
were overriding prefixes in the collection module. This is awkward given we
now have preset rules that look for specific paths in the object key.
Installing the "logwriter" app as a standalone would not work as
expected with a standalone "forwarder" app.

We now "hardcode" the prefixes we require for content type inference.
This ensures users have no way of fudging configuration when using
our apps.