fix: export defaults to environment variables
Setting a default in our env struct currently only affects the value
being read into the struct. It is generally useful to have that value
exporting back to the environment, in the case where other libraries
source their configuration from the environment directly.

An existing usecase is where we modify the default for
`AWS_MAX_ATTEMPTS`. The AWS SDK reads this environment variable
directly when configuring the client.