feat(forwarder): set version on build (#223)
We want to set the version at build time. Unfortunately, the `sam build`
command makes this slightly convoluted, given it executes `make
build-${Resource}` without inheriting any of the variables passed into
our original `make sam-build` call.

As a workaround, we stash the desired version in a file and source it
back in. This is inherently race-y, since concurrent builds may overwite
the file repeatedly. This risk should be fine for the time being, since
we're always writing the same version value and we currently build
serially.