feat: introduce ability to ignore files
We often don't have control of what is writing into a source bucket. As
an example, AWS writes an `aws-programmatic-access-test-object` file
with content-type `application/octet-stream`. Filename inference is not
able to adequately detect that this is a simple text file, and in any
case we don't even want to ingest it.

This commit adds the ability to add a content type override that
disables forwarding for matches files. If a content type is set to
`text/x-ignore`, we will not forward the file.

This is a dubious overloading of the content type field, but it is the
most practical given our configuration constraints. We are currently
surfacing content type overrides consistently in terraform and
cloudformation. Adding a separate variable to capture what files should
be forwarded would add unnecessary complexity to our onboarding.