feat(stack): auto-subscribe to AWS Config
When enabled, AWS Config emits events to EventBridge. This opens the
opportunity to ingest AWS Config data without having to manually
subscribe an S3 bucket (for config history, snapshot and oversized
notifications) and SNS Topic (for change notifications).

We add a new parameter, `ConfigDeliveryBucketName`, which gates whether
to attempt to ingest config data. When present:

- we add the bucket to the source bucket list for our forwarder
- we add an eventbridge rule to construct copy commands for the
  forwarder on successful config delivery
- we add an eventbridge rule to forward all Config events