feat(stack): allow ingesting pre-existing AWS Config
Users with a pre-existing install of AWS Config are currently handed a
DIY project of subscribing their bucket and SNS Topic to our Forwarder.

We can simplify this greatly by allowing users to provide the bucket
where AWS Config delivers data:

- we add this bucket to our "allowed" sources
- we convert eventbridge events for successful data delivery into copy
  events. This ensures we copy Config History and Snapshot files without
  requiring users to explicitly subscribe the bucket.
- we forward change notifications from eventbridge directly, rather than
  relying on SNS Topic subscription.

This functionality will not work for the case where the local bucket is
merely an aggregator for multiple regions / accounts. It also opens the
possibility of ingesting data multiple times, in the case a user
provides both `ConfigDeliveryBucketName` _and_ subscribes that bucket to
our forwarder. Nonetheless it should be a useful feature to have in
place for the base case of getting data in quickly.