feat(stack): auto-subscribe to AWS Config (#233)
* feat(stack): auto-subscribe to AWS Config

When enabled, AWS Config emits events to EventBridge. This opens the
opportunity to ingest AWS Config data without having to manually
subscribe an S3 bucket (for config history, snapshot and oversized
notifications) and SNS Topic (for change notifications).

We add a new parameter, `ConfigDeliveryBucketName`, which gates whether
to attempt to ingest config data. When present:

- we add the bucket to the source bucket list for our forwarder
- we add an eventbridge rule to construct copy commands for the
  forwarder on successful config delivery
- we add an eventbridge rule to forward all Config events

* feat(stack): allow ingesting pre-existing AWS Config

Users with a pre-existing install of AWS Config are currently handed a
DIY project of subscribing their bucket and SNS Topic to our Forwarder.

We can simplify this greatly by allowing users to provide the bucket
where AWS Config delivers data:

- we add this bucket to our "allowed" sources
- we convert eventbridge events for successful data delivery into copy
  events. This ensures we copy Config History and Snapshot files without
  requiring users to explicitly subscribe the bucket.
- we forward change notifications from eventbridge directly, rather than
  relying on SNS Topic subscription.

This functionality will not work for the case where the local bucket is
merely an aggregator for multiple regions / accounts. It also opens the
possibility of ingesting data multiple times, in the case a user
provides both `ConfigDeliveryBucketName` _and_ subscribes that bucket to
our forwarder. Nonetheless it should be a useful feature to have in
place for the base case of getting data in quickly.