fix(forwarder): reduce error wrapping
This commit reduces the amount of spam in error messages derived from
HTTP request failures. Prior to this commit, a 401 would surface as:

```
failed to write messages: failed to process: failed to stream data: failed to consume: failed to upload: %w: unauthorized
```

The `%w` is a bug, while the excessive length can be handled by ditching
error wrapping in some of the chain.