fix(forwarder): ignore empty files on HTTP forwarding
We want to ignore empty files when submitting over HTTP, since they
would produce no records even if the decoder supported an immediate EOF.
Due to a limitation in the AWS SDK, we cannot disambiguate whether a
file size is absent or 0 when processing an S3 bucket notification.
Instead, we inspect the ContentLength of the S3 GetObject response.

This handling is not needed in the Filedrop case, since we handle empty
files in the ingest pipeline, and we only execute a single CopyObject
command.