fix(forwarder): strip path from key for non-s3 destinations
When forwarding data to an S3 destination, the destination "path"
becomes part of the object key. For example, copying `example.json` to
`s3://bucket/file/` should result in a `file/example.json` object key.

When the destination is an HTTP endpoint, the logic differs. Sending the
same file to `https://localhost/endpoint` should not produce a
`endpoint/example.json` key, since `endpoint/` does not pertain to S3 in
any way.

This commit additionally adds tests to our S3 HTTP client. I erroneously
thought the issue was in the `s3http` package, but it turned out to be
due to the forwarder never having considered non-S3 destinations.