fix(forwarder): relax IAM role naming (#285)
The IAM role of the forwarder only needs to be known when forwarding
data to Filedrop. For cases where we are not pointing at an S3 data
access point, we can fallback to letting Cloudformation generating
a unique role name. Doing so avoids the possibility of name clashes when
installing a forwarder in different regions.