fix: introduce otellambda instrumentation (#288)
Wrap our existing mux instrumentation with the `otellambda` package. The
upstream package sets a reasonable span name, kind, and injects
attributes that are consistent with the `faas` semantic convention.

In order to attach the payload to every invocation we still need to
provide our own Invoke method. The payload is added as an attribute.