ci: address release nits
This commit changes the github action we use for semantic releases, and
simplifies the releaserc config file we use.

I changed the action due to maintenance activity. The previous action
has not kept up with dependencies.

For configuration, we're falling back to a default branch configuration.
semantic-release` has a
[reasonable
default](https://github.com/cycjimmy/semantic-release-action?tab=readme-ov-file#branches)
which appears to be consistent across actions. We might as well adopt
that convention in the absence of strong opinions.

Additional nits:
- remove unused releaserc.json
- fix URLs in README.md
- address GHA lint issues