fix(config): adjust precedence of parameters (#296)
The AWS Config API does not allow both exclusion and inclusion sets. We
currently allow users to provide both, and prefer
`ExcludeResourceTypes`, which can be misleading if the intent was to
restrict collection further.

This commit changes precendence to prefer `IncludeResourceTypes` over
`ExcludeResourceTypes`. The reasoning here is that the inclusion list
tends to be small, in which case it is the cost conservative behavior.

An alternative approach would be to assert that only one of the sets is
provided. Unfortunately errors in cloudformation are hard to debug, so
the user experience would be terrible.