fix(forwarder): reduce memory usage
This commit modifies how the buffer queue manages chunks in order to
reduce memory consumption.

Previously, we were appending uncompressed data to a single buffer. On
flush, we would copy the data out and process it in a worker. The worker
would be responsible for compressing the data and posting the data over
HTTP.

This commit applies two improvements:
- compress data on write to queue. This drastically reduces memory
  consumption.
- pass buffer to channel on flush, rather than copying to intermediate
  buffer. There is no justification for copying data from one buffer to
  another.

In a simple test with an 8MB GZIP compressed AWS Config change
notification, our memory consumption went down from ~47MB to ~6MB.