fix(forwarder): cap number of concurrency SQS tasks
The forwarder function processes SQS tasks concurrently. By default this
value was previously unbounded, which would cause significant delay when
redriving a Dead Letter Queue.

This commit changes the default to be set to the runtime number of CPUs,
which allows the behaviour to scale with the amount of memory allocated
to the lambda function.