feat(forwarder): add `SourceObjectKeys` (#308)
This commit adds the ability to specify a set of patterns which identify
object keys we are willing to forward.

A new parameter, `SourceObjectKeys`, allows specifying a set of key
patterns against which incoming objects are filtered. We also use this
parameter to further constrict our eventbridge pattern for triggering
the lambda function.